Split the view.js files
